Устранены замечания после первого ревью. 9 спринт
