--- src/http/BaseHttpHandler.java
@@ -1,22 +1,0 @@
-package http;
-
-import com.sun.net.httpserver.HttpExchange;
-import java.io.IOException;
-import java.nio.charset.StandardCharsets;
-
-public class BaseHttpHandler {
-    protected void sendText(HttpExchange h, String text) throws IOException {
-        byte[] resp = text.getBytes(StandardCharsets.UTF_8);
-        h.getResponseHeaders().add("Content-Type", "application/json;charset=utf-8");
-        h.sendResponseHeaders(200, resp.length);
-        h.getResponseBody().write(resp);
-        h.close();
-    }
-
-    //sendNotFound //для отправки ответа в случае, если объект не был найден;
-
-
-     //для отправки ответа, если при создании или обновлении задача пересекается с уже существующими
-    // sendHasInteractions
-
-}
