--- src/http/HttpTaskManager.java
@@ -1,4 +1,0 @@
-package http;
-
-public class HttpTaskManager {
-}
