+++ src/http/HttpTaskServer.java
@@ -1,13 +1,18 @@
 package http;
 
 import com.google.gson.Gson;
+import com.google.gson.GsonBuilder;
 import com.sun.net.httpserver.HttpServer;
+import http.adapters.DurationAdapter;
+import http.adapters.LocalDateTimeAdapter;
 import http.handlers.*;
 import service.managers.Managers;
 import service.managers.TaskManager;
 
 import java.io.IOException;
 import java.net.InetSocketAddress;
+import java.time.Duration;
+import java.time.LocalDateTime;
 
 public class HttpTaskServer {
 
@@ -28,7 +33,7 @@ public class HttpTaskServer {
 
     public HttpTaskServer(TaskManager taskManager) throws IOException {
         this.taskManager = taskManager;
-        gson = Managers.getGson();
+        gson = getGson();
         server = HttpServer.create(new InetSocketAddress("localhost", PORT), 0);
         server.createContext("/api/v1/tasks", new TasksHandler(this.taskManager, this.gson));
         server.createContext("/api/v1/epics", new EpicsHandler(this.taskManager, this.gson));
@@ -40,6 +45,14 @@ public class HttpTaskServer {
 
     public static void main(String[] args) throws IOException {
         server.start();
+    }
+
+    public static Gson getGson() {
+        GsonBuilder gsonBuilder = new GsonBuilder();
+        gsonBuilder.registerTypeAdapter(LocalDateTime.class, new LocalDateTimeAdapter())
+                .registerTypeAdapter(Duration.class, new DurationAdapter())
+                .serializeNulls();
+        return gsonBuilder.create();
     }
 
     public void stop() {
