+++ src/http/adapters/LocalDateTimeAdapter.java
@@ -10,14 +10,14 @@ import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 
 public class LocalDateTimeAdapter extends TypeAdapter<LocalDateTime> {
-    private static final DateTimeFormatter dtf = DateTimeFormatter.ISO_LOCAL_DATE_TIME;
+    private static final DateTimeFormatter DATE_TIME_FORMATTER = DateTimeFormatter.ISO_LOCAL_DATE_TIME;
 
     @Override
     public void write(final JsonWriter jsonWriter, final LocalDateTime localDateTime) throws IOException {
         if (localDateTime == null) {
             jsonWriter.nullValue();
         } else {
-            jsonWriter.value(localDateTime.format(dtf));
+            jsonWriter.value(localDateTime.format(DATE_TIME_FORMATTER));
         }
     }
 
@@ -27,108 +27,6 @@ public class LocalDateTimeAdapter extends TypeAdapter<LocalDateTime> {
             jsonReader.nextNull();
             return null;
         }
-        return LocalDateTime.parse(jsonReader.nextString(), dtf);
+        return LocalDateTime.parse(jsonReader.nextString(), DATE_TIME_FORMATTER);
     }
 }
-
-
-
-
-
-
-
-
-
-/*
-public class LocalDateTimeAdapter extends TypeAdapter<LocalDateTime> {
-    private static final DateTimeFormatter formatter = DateTimeFormatter.ISO_LOCAL_DATE_TIME;
-
-    @Override
-    public void write(JsonWriter out, LocalDateTime value) throws IOException {
-        if (value == null) {
-            out.nullValue();
-        } else {
-            out.value(value.format(formatter));
-        }
-    }
-
-    @Override
-    public LocalDateTime read(JsonReader in) throws IOException {
-        if (in.peek() == null) {
-            in.nextNull();
-            return null;
-        } else {
-            return LocalDateTime.parse(in.nextString(), formatter);
-        }
-    }
-}
-
-
-
-
-
-
-
-/*
- public static class LocalDateAdapter implements JsonSerializer<LocalDate>, JsonDeserializer<LocalDate> {
-
-        private final DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd");
-
-        @Override
-        public JsonElement serialize(final LocalDate date, final Type typeOfSrc,
-                                     final JsonSerializationContext context) {
-            return new JsonPrimitive(date.format(formatter));
-        }
-
-        @Override
-        public LocalDate deserialize(final JsonElement json, final Type typeOfT,
-                                     final JsonDeserializationContext context) throws JsonParseException {
-            return LocalDate.parse(json.getAsString(), formatter);
-        }
-    }
-    public static class DurationAdapter extends TypeAdapter<Duration> {
-
-        @Override
-        public void write(JsonWriter out, Duration value) throws IOException {
-            if (value == null) {
-                out.nullValue();
-            } else {
-                out.value(value.toString());
-            }
-        }
-
-
-        @Override
-        public Duration read(JsonReader in) throws IOException {
-            if (in.peek() == null) {
-                in.nextNull();
-                return null;
-            } else {
-                return Duration.parse(in.nextString());
-            }
-        }
-
-    }
-    public static class LocalDateTimeAdapter extends TypeAdapter<LocalDateTime> {
-        private static final DateTimeFormatter formatter = DateTimeFormatter.ISO_LOCAL_DATE_TIME;
-
-        @Override
-        public void write(JsonWriter out, LocalDateTime value) throws IOException {
-            if (value == null) {
-                out.nullValue();
-            } else {
-                out.value(value.format(formatter));
-            }
-        }
-
-        @Override
-        public LocalDateTime read(JsonReader in) throws IOException {
-            if (in.peek() == null) {
-                in.nextNull();
-                return null;
-            } else {
-                return LocalDateTime.parse(in.nextString(), formatter);
-            }
-        }
-    }
-*/
